+++ lib/mint_flutter_extensions.dart
@@ -1,3 +1,5 @@
 library mint_flutter_extensions;
 
 export 'src/build_context_colors_extensions.dart';
+export 'src/build_context_text_styles_extensions.dart';
+export 'src/text_style_extensions.dart';